Updated reinstall cadpage logic to use available system API

+++ src/main/java/net/anei/cadpage/donation/ClearDataEvent.java
@@ -1,10 +1,13 @@
 package net.anei.cadpage.donation;
 
 import android.app.Activity;
+import android.app.ActivityManager;
 import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
+import android.os.Build;
+
 import net.anei.cadpage.CadPageActivity;
 import net.anei.cadpage.ManagePreferences;
 import net.anei.cadpage.R;
@@ -21,26 +24,36 @@ public class ClearDataEvent extends DonateEvent {
 
   @Override
   protected void doEvent(Activity activity) {
-    
-    // Delete any private files
-    String[] files = activity.fileList();
-    for (String file : files) activity.deleteFile(file);
-    
-    // Clear any shared preferences
-    ManagePreferences.clearAll();
-    VendorManager.instance().clearAll();
-    
-    // Set up an intent to relaunch Cadpage in 1 second
-    AlarmManager myAM = (AlarmManager) activity.getSystemService(Context.ALARM_SERVICE);
-    Intent intent = CadPageActivity.getLaunchIntent(activity);
-    PendingIntent pendIntent =
-      PendingIntent.getActivity(activity, 0, intent, PendingIntent.FLAG_ONE_SHOT);
 
-    long triggerTime = System.currentTimeMillis() + 1000L;
-    myAM.set(AlarmManager.RTC_WAKEUP, triggerTime, pendIntent);
+    // They gave us a nifty API to do this in Android 4.4
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+      ActivityManager fm = (ActivityManager)activity.getSystemService(Context.ACTIVITY_SERVICE);
+      assert fm != null;
+      fm.clearApplicationUserData();
+    }
 
-    // Kill off this process and let it relaunch
-    System.exit(2);
+    // Have to do this the old fashion way
+    else {
+      // Delete any private files
+      String[] files = activity.fileList();
+      for (String file : files) activity.deleteFile(file);
+
+      // Clear any shared preferences
+      ManagePreferences.clearAll();
+      VendorManager.instance().clearAll();
+
+      // Set up an intent to relaunch Cadpage in 1 second
+      AlarmManager myAM = (AlarmManager) activity.getSystemService(Context.ALARM_SERVICE);
+      Intent intent = CadPageActivity.getLaunchIntent(activity);
+      PendingIntent pendIntent =
+              PendingIntent.getActivity(activity, 0, intent, PendingIntent.FLAG_ONE_SHOT);
+
+      long triggerTime = System.currentTimeMillis() + 1000L;
+      myAM.set(AlarmManager.RTC_WAKEUP, triggerTime, pendIntent);
+
+      // Kill off this process and let it relaunch
+      System.exit(2);
+    }
   }
   
   private static final ClearDataEvent instance = new ClearDataEvent();
